requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.0.4
atomicwrites==1.1.5
attrs==18.1.0
Automat==0.7.0
autopep8==1.3.5
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.11.29
cffi==1.11.5
colorama==0.3.9
constantly==15.1.0
cryptography==2.4.2
cryptography-vectors==2.3
cssselect==1.0.3
cycler==0.10.0
fake-useragent==0.1.10
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
isort==4.3.4
json2table==1.1.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.2.4
mccabe==0.6.1
mkl-fft==1.0.6
mkl-random==1.0.1
more-itertools==4.3.0
olefile==0.46
parsel==1.4.0
Pillow==5.2.0
pluggy==0.7.1
py==1.5.4
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycparser==2.18
PyDispatcher==2.0.5
pylint==2.1.1
pyparsing==2.2.2
PySocks==1.6.8
pytesseract==0.2.6
pytest==3.7.1
pytest-runner==4.2
python-dateutil==2.7.5
pytz==2018.9
pywin32==223
queuelib==1.5.0
six==1.11.0
w3lib==1.19.0
wincertstore==0.2
wrapt==1.10.11
zope.interface==4.5.0
